Replaces bin/run -> Makefile and requirements.txt -> setup.py
`setup.py` is a dependency tracker on its own and recursively resolves.
It's less complex this way.

Make is a dev toolset only intended to supplement. The documentation
makes no mention of it on purpose, only sharing the implementation of
the Makefile as if it wasn't even there. It only exists to make
development easier by grouping common tasks together.