Consistent docs, project structure, and prep for cloud hosting (#2)
* Consistent README docs across projects

* fixup! Consistent README docs across projects

* Consistent setup.py structure

* Replaces bin/run -> Makefile and requirements.txt -> setup.py

`setup.py` is a dependency tracker on its own and recursively resolves.
It's less complex this way.

Make is a dev toolset only intended to supplement. The documentation
makes no mention of it on purpose, only sharing the implementation of
the Makefile as if it wasn't even there. It only exists to make
development easier by grouping common tasks together.

* License should refer to an entity that actually exists

* Default reviewers for each code section

* Setup.py in consistent format and pruned dependencies

* Adds Travis-CI manifest

* Adds all the badges

* Adds basic test

* Fixes test since command -v doesn't exist on TravisCI

* Ignore vulnerabilities in test directory

* Removes trailing whitespace

* Dives all-in to codacy