Variable shadowing is bad and you should feel bad
Just kidding. It is bad though. We should be using only what is passed
into the method instead of some global state store like that, unless we
specifically know that is what we are doing. Even then, local passed-in
variables are better. The moment the arg is yanked, variable shadowing
defers to the previously defined variable value, which would the the
`config` global. That may not be a bad thing, but it is another level of
indrection that makes troubleshooting problems far less obvious.